Fix the description with the documentation generator
Make sure we ignore the patched classes and fix an issue with the rspec
suite.
